requirements.txt
absl-py>=0.8
thrift==0.11.0